# Remove extra copy of inline strings in CAST(DECIMAL as VARCHAR)