This PR writes inline strings directly to the vector's values buffer. This 
buffer has 16 bytes slots, one per value. First 4 bytes store string size, the 
remaining 12 bytes store the string. In this way, extra copy of inline strings 
can be removed. Benchmark shows close performance after this change.

Benchmark  | Before | After
-- | -- | --
cast_decimal_to_inline_string | 8.29ms | 8.24ms