@rui-mo 2023-09-25
@mbasmanova I tried to enlarge the vector size of benchmark as `100'000'000`, and below was the benchmark result. Is the improvement limited with the consideration of additional complexity? If so, we can close this PR.

Benchmark  | Before | After
-- | -- | --
cast_decimal_to_inline_string | 22.49s | 21.67s