@mbasmanova 2023-09-26
@rui-mo Rui, thank you for benchmarking. 100M row vectors seem not very practical. I'd say we can close this PR since the improvement is not significant.